requirements.txt
django
pillow
django-rest-framework